backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.8.2
openai==1.51.0
qdrant-client==1.11.3
python-dotenv==1.0.1
httpx==0.27.2
slowapi==0.1.9
